Controlador, terminar

+++ src/main/java/com/illuminati/ebs/repository/TipoEntregaPedidoRepository.java
@@ -4,5 +4,5 @@ import com.illuminati.ebs.entity.TipoEntregaPedido;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoEntregaPedidoRepository extends GenericRepository<TipoEntregaPedido, Integer>{
+public interface TipoEntregaPedidoRepository extends GenericRepository<TipoEntregaPedido, Long>{
 }
