+++ src/main/java/com/illuminati/ebs/repository/UnidadMedidaRepository.java
@@ -4,5 +4,5 @@ import com.illuminati.ebs.entity.UnidadMedida;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UnidadMedidaRepository extends GenericRepository<UnidadMedida, Integer>{
+public interface UnidadMedidaRepository extends GenericRepository<UnidadMedida, Long>{
 }
